chore: rename tanstack operations
